+++ src/main/java/org/jabref/logic/citationstyle/CitationStylePreviewLayout.java
@@ -5,7 +5,7 @@ import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
 
 public class CitationStylePreviewLayout implements PreviewLayout {
-    private CitationStyle citationStyle;
+    private final CitationStyle citationStyle;
 
     public CitationStylePreviewLayout(CitationStyle citationStyle) {
         this.citationStyle = citationStyle;
@@ -27,5 +27,10 @@ public class CitationStylePreviewLayout implements PreviewLayout {
 
     public String getFilePath() {
         return citationStyle.getFilePath();
+    }
+
+    @Override
+    public String getInternalName() {
+        return citationStyle.getTitle();
     }
 }