+++ src/main/java/org/jabref/logic/preview/PreviewLayout.java
@@ -11,4 +11,6 @@ public interface PreviewLayout {
     String generatePreview(BibEntry entry, BibDatabase database);
 
     String getName();
+
+    String getInternalName();
 }